pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-session-delete"
version = "1.0.4"
description = "External Codex App session delete injector"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31",
  "websocket-client>=1.7",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[project.scripts]
codex-session-delete = "codex_session_delete.cli:main"

[tool.setuptools.package-data]
codex_session_delete = ["assets/*", "inject/*.js"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
